linkedin-worker/requirements.txt
playwright==1.40.0
playwright-stealth==2.1.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
